only GOD understands the errors now
Just removed the derive Defaults and defaulted to using own self made
default implementations even though there is literally ZERO reason for
it to fail as the types used are either have default null or are
optional.

Also fixed some borrowing/mutability errors, had to use a for loop
like a caveman in one place.

Fiddled with knobs and made them into references.

Tried adding the marker field and still failed (I think because of
that.)

Now going to remove the marker field again and trying to see if it works