Left the safety of code at the mercy of carefulness and GOD
I just gave up trying to make the factory 'always produces the type of
object with which it was initialized'.
Now each usage of the method produce will require you to pass in the
type of the object being produced.
Worse of all, you might accidentally create a factory with a type T in
mind (and code) but when producing you specify type U.

It was all because for some reason the rust compiler couldnot let go
of the 'Sized' '?Sized' debacle for some reason.

Might goto Box, dyn, Rc, things now, but I dont give a f anymore.