Fixed the not dropping of vulkan resources before quitting errors
Yet again got a 'map is lazy' error that was actually not dropping the
allocated device arrays. For one part used a standard for in range
loop, for another part, used a proper map().collect() mechanism.
Added a few Debug derivations for debugging the whole thing. (I cannot
believe that each child object has a reference to the whole thing upto
the parent for not much reason by now).

Also added the destroy for the debug messenger (that i had forgotten
to add) in the cases of ![cfg(debug_assertions)]